+++ packages/shared/src/components/post/common/PostContentReminder.tsx
@@ -1,16 +1,10 @@
 import React, { ReactElement } from 'react';
 import { PostContentWidget } from './PostContentWidget';
-import {
-  ReminderPreference,
-  useBookmarkReminder,
-  useBookmarkReminderEnrollment,
-} from '../../../hooks/notifications';
+import { useBookmarkReminderEnrollment } from '../../../hooks/notifications';
 import { Post } from '../../../graphql/posts';
 import { BookmarkReminderIcon } from '../../icons/Bookmark/Reminder';
-import { Button, ButtonSize, ButtonVariant } from '../../buttons/Button';
 import { IconSize } from '../../Icon';
-import { useLazyModal } from '../../../hooks/useLazyModal';
-import { LazyModal } from '../../modals/common/types';
+import { PostReminderOptions } from './PostReminderOptions';
 
 interface PostContentReminderProps {
   post: Post;
@@ -20,19 +14,10 @@ export function PostContentReminder({
   post,
 }: PostContentReminderProps): ReactElement {
   const shouldShowReminder = useBookmarkReminderEnrollment(post);
-  const { onBookmarkReminder } = useBookmarkReminder();
-  const { openModal } = useLazyModal();
 
   if (!shouldShowReminder) {
     return null;
   }
-
-  const runBookmarkReminder = (preference: ReminderPreference) =>
-    onBookmarkReminder({
-      postId: post.id,
-      existingReminder: post.bookmark?.remindAt,
-      preference,
-    });
 
   return (
     <PostContentWidget
@@ -40,29 +25,7 @@ export function PostContentReminder({
       icon={<BookmarkReminderIcon size={IconSize.Small} />}
       title="Want to read / watch it later?"
     >
-      <span className="ml-auto flex flex-row gap-3">
-        <Button
-          variant={ButtonVariant.Float}
-          size={ButtonSize.XSmall}
-          onClick={() => runBookmarkReminder(ReminderPreference.OneHour)}
-        >
-          1h
-        </Button>
-        <Button
-          variant={ButtonVariant.Float}
-          size={ButtonSize.XSmall}
-          onClick={() => runBookmarkReminder(ReminderPreference.Tomorrow)}
-        >
-          24h
-        </Button>
-        <Button
-          variant={ButtonVariant.Float}
-          size={ButtonSize.XSmall}
-          onClick={() => openModal({ type: LazyModal.ReadItLater })}
-        >
-          Other
-        </Button>
-      </span>
+      <PostReminderOptions post={post} className="ml-auto" />
     </PostContentWidget>
   );
 }
